requirements.txt
streamlit
langchain-community
langchain-openai
langchain
langgraph
pymupdf
faiss-cpu
streamlit-pdf-viewer
python-dotenv
pylint
isort
ruff